Removed repetition in export methods
Improved export methods in Exam and Section. I wrote a function which adds a Model object's attribute to a dict only if it is not equal to the default value of the model's field. So I loop over an object's relevant attributes and add them to the dict using that function. This will also make it easier to change a field's default value.
Added blank=True in fields with null=True. This resulted in a new migration.